+++ packages/client/src/pages/Stats/EntitiesTab/EntitiesTab.tsx
@@ -5,7 +5,7 @@ import api from "api";
 import { Button, ButtonGroup, Input, Loader, Timestamp } from "components";
 import { useResizeObserver } from "hooks";
 import React, { useEffect, useMemo, useReducer, useState } from "react";
-import { FaCalendarPlus, FaTimes } from "react-icons/fa";
+import { FaCalendarPlus, FaDatabase, FaSyncAlt, FaTimes } from "react-icons/fa";
 import { toast } from "react-toastify";
 import {
   StyledDateInputWrapper,
@@ -21,6 +21,7 @@ import { applyUserThreshold, datePickerToIso, isoToDatePicker } from "../utils";
 import { StatsChart } from "./StatsChart/StatsChart";
 import { StatsTable } from "./StatsTable/StatsTable";
 import { USER_THRESHOLD_MAX } from "../constants";
+import { AttributeButtonGroup } from "components/advanced/AttributeButtonGroup/AttributeButtonGroup";
 
 export const EntitiesTab: React.FC = () => {
   const [state, dispatch] = useReducer(statsReducer, initialState);
@@ -143,232 +144,282 @@ export const EntitiesTab: React.FC = () => {
     }
   }, [isError]);
 
+  // get user data
+  const userId = localStorage.getItem("userid");
+  const { data: user } = useQuery({
+    queryKey: ["user", userId],
+    queryFn: async () => {
+      const res = await api.usersGet(userId as string);
+      return res.data ?? undefined;
+    },
+    enabled: !!userId && api.isLoggedIn(),
+  });
+
+  const allowMaterializedStats = user?.options.allowMaterializedStats ?? false;
+
   return (
-    <StyledEntitiesLayout>
-      <StyledFieldGroup>
-        {/* Date From */}
-        <StyledField>
-          <StyledFieldLabel>From Date</StyledFieldLabel>
-          {!state.showDateFromRangePicker ? (
-            <StyledDateInputWrapper>
-              <Timestamp label="From" value={state.dateFrom} />
-              <Button
-                icon={<FaCalendarPlus />}
-                onClick={() => {
+    <>
+      {/* {allowMaterializedStats && (
+        <span style={{ display: "flex", zIndex: 30, marginRight: "4rem" }}>
+          <AttributeButtonGroup
+            noMargin
+            options={[
+              {
+                icon: <FaSyncAlt size={10} />,
+                longValue: "Classic (Live Data)",
+                shortValue: "Classic",
+                onClick: () => {
                   dispatch({
-                    type: "showDateFromRangePickerUpdate",
-                    payload: true,
-                  });
-                  dispatch({
-                    type: "dateFromUpdate",
-                    payload: new Date(
-                      new Date().setFullYear(new Date().getFullYear() - 5)
-                    ).toISOString(),
-                  });
-                }}
-                color="primary"
-                inverted
-                tooltipLabel="Add custom date from"
-                noBorder
-                noBackground
-              />
-            </StyledDateInputWrapper>
-          ) : (
-            <StyledDateInputWrapper>
-              <Input
-                type="date"
-                value={isoToDatePicker(state.dateFrom)}
-                onChangeFn={(value) =>
-                  dispatch({
-                    type: "dateFromUpdate",
-                    payload: datePickerToIso(value),
-                  })
-                }
-              />
-              <Button
-                icon={<FaTimes />}
-                onClick={() => {
-                  dispatch({
-                    type: "showDateFromRangePickerUpdate",
+                    type: "useMaterializedUpdate",
                     payload: false,
                   });
+                },
+                selected: !state.useMaterialized,
+              },
+              {
+                icon: <FaDatabase />,
+                longValue: "Fast (Pre-calculated)",
+                shortValue: "Fast",
+                onClick: () => {
                   dispatch({
-                    type: "dateFromUpdate",
-                    payload: new Date("2000-01-01").toISOString(),
-                  });
-                }}
-                color="primary"
-                inverted
-                tooltipLabel="Reset to Since Forever"
-                noBackground
-              />
-            </StyledDateInputWrapper>
-          )}
-        </StyledField>
-
-        {/* Date To */}
-        <StyledField>
-          <StyledFieldLabel>To Date</StyledFieldLabel>
-          {!state.showDateToRangePicker ? (
-            <StyledDateInputWrapper>
-              <Timestamp label="To" value={state.dateTo} />
-              <Button
-                icon={<FaCalendarPlus />}
-                onClick={() => {
-                  dispatch({
-                    type: "dateToUpdate",
-                    payload: new Date().toISOString(),
-                  });
-                  dispatch({
-                    type: "showDateToRangePickerUpdate",
+                    type: "useMaterializedUpdate",
                     payload: true,
                   });
-                }}
-                color="primary"
-                inverted
-                tooltipLabel="Add custom date to"
-                noBorder
-                noBackground
-              />
-            </StyledDateInputWrapper>
-          ) : (
-            <StyledDateInputWrapper>
-              <Input
-                type="date"
-                value={isoToDatePicker(state.dateTo)}
-                onChangeFn={(value) =>
-                  dispatch({
-                    type: "dateToUpdate",
-                    payload: datePickerToIso(value),
-                  })
-                }
-              />
-              <Button
-                icon={<FaTimes />}
-                onClick={() => {
-                  dispatch({
-                    type: "showDateToRangePickerUpdate",
-                    payload: false,
-                  });
-                  updateToCurrentTime();
-                }}
-                color="primary"
-                inverted
-                tooltipLabel="Reset to Until Now"
-                noBackground
-              />
-            </StyledDateInputWrapper>
-          )}
-        </StyledField>
-
-        <StyledField>
-          <StyledFieldLabel>Time Unit</StyledFieldLabel>
-          <ButtonGroup $noMarginRight>
-            {Object.values(TimeUnit).map((unit) => (
-              <Button
-                key={unit}
-                label={String(unit)}
-                onClick={() => {
-                  dispatch({
-                    type: "timeUnitUpdate",
-                    payload: unit as TimeUnit,
-                  });
-                }}
-                color={state.timeUnit === unit ? "primary" : "grey"}
-              />
-            ))}
-          </ButtonGroup>
-        </StyledField>
-
-        <StyledField>
-          <StyledFieldLabel>Event type</StyledFieldLabel>
-          <ButtonGroup $noMarginRight>
-            {Object.values(EventType).map((eventType) => (
-              <Button
-                key={eventType}
-                label={String(eventType)}
-                onClick={() => {
-                  dispatch({
-                    type: "eventTypeUpdate",
-                    payload: eventType,
-                  });
-                }}
-                color={state.eventType.includes(eventType) ? "primary" : "grey"}
-              />
-            ))}
-          </ButtonGroup>
-        </StyledField>
-
-        <StyledField>
-          <StyledFieldLabel>Aggregate By</StyledFieldLabel>
-          <ButtonGroup $noMarginRight>
-            {Object.values(Aggregation).map((agg) => (
-              <Button
-                key={agg}
-                label={String(agg)}
-                onClick={() => {
-                  dispatch({ type: "aggregateUpdate", payload: agg });
-                }}
-                color={state.aggregate === agg ? "primary" : "grey"}
-              />
-            ))}
-          </ButtonGroup>
-        </StyledField>
-        {state.aggregate === Aggregation.USER && (
+                },
+                selected: state.useMaterialized,
+              },
+            ]}
+          />
+        </span>
+      )} */}
+      <StyledEntitiesLayout>
+        <StyledFieldGroup>
+          {/* Date From */}
           <StyledField>
-            <StyledFieldLabel>Ignore users below %</StyledFieldLabel>
-            <Input
-              type="number"
-              value={String(usersIgnoreBelowValue)}
-              onChangeFn={(value) => {
-                const num = Number(value);
-                const safe = Math.min(
-                  USER_THRESHOLD_MAX,
-                  Math.max(0, Number.isFinite(num) ? num : 0)
-                );
-                setUsersIgnoreBelowValue(safe);
-              }}
-              changeOnType
-              min={0}
-              max={USER_THRESHOLD_MAX}
-            />
+            <StyledFieldLabel>From Date</StyledFieldLabel>
+            {!state.showDateFromRangePicker ? (
+              <StyledDateInputWrapper>
+                <Timestamp label="From" value={state.dateFrom} />
+                <Button
+                  icon={<FaCalendarPlus />}
+                  onClick={() => {
+                    dispatch({
+                      type: "showDateFromRangePickerUpdate",
+                      payload: true,
+                    });
+                    dispatch({
+                      type: "dateFromUpdate",
+                      payload: new Date(
+                        new Date().setFullYear(new Date().getFullYear() - 5)
+                      ).toISOString(),
+                    });
+                  }}
+                  color="primary"
+                  inverted
+                  tooltipLabel="Add custom date from"
+                  noBorder
+                  noBackground
+                />
+              </StyledDateInputWrapper>
+            ) : (
+              <StyledDateInputWrapper>
+                <Input
+                  type="date"
+                  value={isoToDatePicker(state.dateFrom)}
+                  onChangeFn={(value) =>
+                    dispatch({
+                      type: "dateFromUpdate",
+                      payload: datePickerToIso(value),
+                    })
+                  }
+                />
+                <Button
+                  icon={<FaTimes />}
+                  onClick={() => {
+                    dispatch({
+                      type: "showDateFromRangePickerUpdate",
+                      payload: false,
+                    });
+                    dispatch({
+                      type: "dateFromUpdate",
+                      payload: new Date("2000-01-01").toISOString(),
+                    });
+                  }}
+                  color="primary"
+                  inverted
+                  tooltipLabel="Reset to Since Forever"
+                  noBackground
+                />
+              </StyledDateInputWrapper>
+            )}
           </StyledField>
+
+          {/* Date To */}
+          <StyledField>
+            <StyledFieldLabel>To Date</StyledFieldLabel>
+            {!state.showDateToRangePicker ? (
+              <StyledDateInputWrapper>
+                <Timestamp label="To" value={state.dateTo} />
+                <Button
+                  icon={<FaCalendarPlus />}
+                  onClick={() => {
+                    dispatch({
+                      type: "dateToUpdate",
+                      payload: new Date().toISOString(),
+                    });
+                    dispatch({
+                      type: "showDateToRangePickerUpdate",
+                      payload: true,
+                    });
+                  }}
+                  color="primary"
+                  inverted
+                  tooltipLabel="Add custom date to"
+                  noBorder
+                  noBackground
+                />
+              </StyledDateInputWrapper>
+            ) : (
+              <StyledDateInputWrapper>
+                <Input
+                  type="date"
+                  value={isoToDatePicker(state.dateTo)}
+                  onChangeFn={(value) =>
+                    dispatch({
+                      type: "dateToUpdate",
+                      payload: datePickerToIso(value),
+                    })
+                  }
+                />
+                <Button
+                  icon={<FaTimes />}
+                  onClick={() => {
+                    dispatch({
+                      type: "showDateToRangePickerUpdate",
+                      payload: false,
+                    });
+                    updateToCurrentTime();
+                  }}
+                  color="primary"
+                  inverted
+                  tooltipLabel="Reset to Until Now"
+                  noBackground
+                />
+              </StyledDateInputWrapper>
+            )}
+          </StyledField>
+
+          <StyledField>
+            <StyledFieldLabel>Time Unit</StyledFieldLabel>
+            <ButtonGroup $noMarginRight>
+              {Object.values(TimeUnit).map((unit) => (
+                <Button
+                  key={unit}
+                  label={String(unit)}
+                  onClick={() => {
+                    dispatch({
+                      type: "timeUnitUpdate",
+                      payload: unit as TimeUnit,
+                    });
+                  }}
+                  color={state.timeUnit === unit ? "primary" : "grey"}
+                />
+              ))}
+            </ButtonGroup>
+          </StyledField>
+
+          <StyledField>
+            <StyledFieldLabel>Event type</StyledFieldLabel>
+            <ButtonGroup $noMarginRight>
+              {Object.values(EventType).map((eventType) => (
+                <Button
+                  key={eventType}
+                  label={String(eventType)}
+                  onClick={() => {
+                    dispatch({
+                      type: "eventTypeUpdate",
+                      payload: eventType,
+                    });
+                  }}
+                  color={
+                    state.eventType.includes(eventType) ? "primary" : "grey"
+                  }
+                />
+              ))}
+            </ButtonGroup>
+          </StyledField>
+
+          <StyledField>
+            <StyledFieldLabel>Aggregate By</StyledFieldLabel>
+            <ButtonGroup $noMarginRight>
+              {Object.values(Aggregation).map((agg) => (
+                <Button
+                  key={agg}
+                  label={String(agg)}
+                  onClick={() => {
+                    dispatch({ type: "aggregateUpdate", payload: agg });
+                  }}
+                  color={state.aggregate === agg ? "primary" : "grey"}
+                />
+              ))}
+            </ButtonGroup>
+          </StyledField>
+          {state.aggregate === Aggregation.USER && (
+            <StyledField>
+              <StyledFieldLabel>Ignore users below %</StyledFieldLabel>
+              <Input
+                type="number"
+                value={String(usersIgnoreBelowValue)}
+                onChangeFn={(value) => {
+                  const num = Number(value);
+                  const safe = Math.min(
+                    USER_THRESHOLD_MAX,
+                    Math.max(0, Number.isFinite(num) ? num : 0)
+                  );
+                  setUsersIgnoreBelowValue(safe);
+                }}
+                changeOnType
+                min={0}
+                max={USER_THRESHOLD_MAX}
+              />
+            </StyledField>
+          )}
+
+          <Button
+            color="success"
+            label={state.useMaterialized ? "Aggregate" : "Refresh"}
+            disabled={isLoadingStats || isAggregating}
+            onClick={
+              state.useMaterialized
+                ? () => void aggregateMutateAsync()
+                : updateToCurrentTime
+            }
+          />
+        </StyledFieldGroup>
+
+        {data && (
+          <>
+            <StyledResultsChart ref={chartRef}>
+              <StatsChart
+                data={data}
+                height={chartHeight ? Math.max(0, chartHeight) : 0}
+                width={chartWidth ? Math.max(0, chartWidth - 50) : 0}
+                request={statsRequest}
+              />
+            </StyledResultsChart>
+            <StyledResultsTable ref={tableRef}>
+              <StatsTable
+                data={data}
+                height={tableHeight ? Math.max(0, tableHeight) : 0}
+                width={tableWidth ? Math.max(0, tableWidth - 50) : 0}
+                request={statsRequest}
+              />
+            </StyledResultsTable>
+          </>
         )}
 
-        <Button
-          color="success"
-          label={state.useMaterialized ? "Aggregate" : "Refresh"}
-          disabled={isLoadingStats || isAggregating}
-          onClick={
-            state.useMaterialized
-              ? () => void aggregateMutateAsync()
-              : updateToCurrentTime
-          }
-        />
-      </StyledFieldGroup>
-
-      {data && (
-        <>
-          <StyledResultsChart ref={chartRef}>
-            <StatsChart
-              data={data}
-              height={chartHeight ? Math.max(0, chartHeight) : 0}
-              width={chartWidth ? Math.max(0, chartWidth - 50) : 0}
-              request={statsRequest}
-            />
-          </StyledResultsChart>
-          <StyledResultsTable ref={tableRef}>
-            <StatsTable
-              data={data}
-              height={tableHeight ? Math.max(0, tableHeight) : 0}
-              width={tableWidth ? Math.max(0, tableWidth - 50) : 0}
-              request={statsRequest}
-            />
-          </StyledResultsTable>
-        </>
-      )}
-
-      <Loader show={isLoadingStats || isAggregating} />
-    </StyledEntitiesLayout>
+        <Loader show={isLoadingStats || isAggregating} />
+      </StyledEntitiesLayout>
+    </>
   );
 };
